+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# On dev - *unreleased*
+
+## Added
+
+- IDErioja.Relieve
+- IDErioja.Claro
+- IDErioja.Oscuro
+
+## Changed
+- IDErioja now is IDErioja.Base 
+
+
+### Checks
+
+- [ ] Bump version on script
+- [ ] Beautify
+- [ ] Lint
+
 # v1.3.2 - *2022-11-14*
 
 Cleanup